requirements.txt
bs4==0.0.1
python-twitter==3.5
requests==2.22.0
